# openapi-integration.yaml
openapi: 3.0.3
info:
  title: VSMS Integration API
  description: Integration endpoints for cross-system queries. Authenticated via X-Api-Key header.
  version: 1.0.0

servers:
  - url: http://localhost:3001/api/integration
    description: Local development

security:
  - apiKey: []

components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-Api-Key

  schemas:
    ProgressResults:
      type: object
      properties:
        pass:             { type: integer }
        fail:             { type: integer }
        conditional_pass: { type: integer }
        blocked:          { type: integer }
        not_tested:       { type: integer }
        not_applicable:   { type: integer }

    VtmsProgress:
      type: object
      nullable: true
      properties:
        planId:          { type: string }
        planName:        { type: string }
        planStatus:      { type: string }
        totalItems:      { type: integer }
        latestRunStatus: { type: string, nullable: true }
        results:         { $ref: '#/components/schemas/ProgressResults' }
        completionPct:   { type: integer }

    Schedule:
      type: object
      properties:
        id:              { type: string }
        projectName:     { type: string }
        taskDescription: { type: string }
        testEngineer:    { type: string }
        testUnit:        { type: string, nullable: true }
        startDate:       { type: string }
        endDate:         { type: string }
        isCompleted:     { type: boolean }
        isDelayed:       { type: boolean }
        isCancelled:     { type: boolean, description: 'True if cancelled in VSMS; mutually exclusive with isCompleted' }
        delayReason:     { type: string, nullable: true }
        vtmsPlanId:      { type: string, nullable: true }
        completedAt:     { type: string, format: date-time, nullable: true, description: 'Timestamp when isCompleted last transitioned to true' }
        userFlag:        { type: boolean, description: 'User (non-admin) mark 標記' }
        userFlagNote:    { type: string, nullable: true, description: 'Note attached to the user mark' }

    ScheduleWithProgress:
      allOf:
        - $ref: '#/components/schemas/Schedule'
        - type: object
          properties:
            vtmsProgress:
              $ref: '#/components/schemas/VtmsProgress'

paths:
  /schedules:
    get:
      summary: List schedules with optional filters
      parameters:
        - { in: query, name: testUnit,     schema: { type: string }, description: 'Exact test unit name' }
        - { in: query, name: testEngineer, schema: { type: string }, description: 'Exact test engineer name' }
        - { in: query, name: device,       schema: { type: string }, description: 'Exact equipment/device name' }
        - { in: query, name: projectName,  schema: { type: string }, description: 'Partial project name match' }
        - { in: query, name: isCompleted,  schema: { type: string, enum: ['true','false'] } }
        - { in: query, name: isDelayed,    schema: { type: string, enum: ['true','false'] } }
        - { in: query, name: isCancelled,  schema: { type: string, enum: ['true','false'] } }
        - { in: query, name: userFlag,     schema: { type: string, enum: ['true','false'] }, description: 'Filter by user (non-admin) mark' }
        - { in: query, name: dateFrom,     schema: { type: string, description: 'YYYY-MM-DD (matches startDate >=)' } }
        - { in: query, name: dateTo,       schema: { type: string, description: 'YYYY-MM-DD (matches startDate <=)' } }
      responses:
        '200':
          description: Schedule array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schedule'
        '401':
          description: Unauthorized

  /schedules/summary:
    get:
      summary: Schedule health summary by unit
      responses:
        '200':
          description: Summary statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:       { type: integer }
                  completed:   { type: integer, description: 'Mutually exclusive bucket; excludes cancelled schedules' }
                  delayed:     { type: integer, description: 'Mutually exclusive bucket; excludes cancelled schedules' }
                  inProgress:  { type: integer, description: 'Mutually exclusive bucket; excludes cancelled schedules' }
                  notStarted:  { type: integer, description: 'Mutually exclusive bucket; excludes cancelled schedules' }
                  cancelled:   { type: integer, description: 'Schedules cancelled in VSMS; not counted in the other buckets' }
                  byUnit:
                    type: object
                    additionalProperties: { type: integer }
        '401':
          description: Unauthorized

  /engineers:
    get:
      summary: Resolve a loose engineer name to the canonical testEngineer value
      description: >
        Users refer to engineers informally ("darius", "DARIUS", "darius chang") but
        testEngineer is stored as 'Darius_Chang'. Matching is deterministic and lives
        on the server so callers never guess a name — a wrong guess silently returns
        another person's data (e.g. "will" overlaps Will_Wang / William_Wu / Willie_Lin).
        Results are the roster (engineers table) unioned with any name that appears in
        schedules, so historical engineers stay resolvable.
      parameters:
        - in: query
          name: q
          required: false
          schema: { type: string }
          description: >
            How the user referred to the engineer. Case-insensitive; underscore, space
            and hyphen are equivalent separators. Omit to list every engineer.
      responses:
        '200':
          description: >
            Candidates, strongest match first then by name. An exact full-name match is
            returned alone; otherwise name/surname/prefix matches are returned together
            so the caller can disambiguate. Empty array means no match — do not retry
            with a different guess.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:     { type: string, description: 'Canonical value; use verbatim as testEngineer' }
                    testUnit: { type: string, nullable: true }
                    isActive: { type: boolean, description: 'false = has left; still appears in past schedules' }
                    matchType:
                      type: string
                      nullable: true
                      enum: [exact, firstName, lastName, prefix, contains]
                      description: 'Match strength; null when q was omitted'
        '401':
          description: Unauthorized

  /schedules/by-plan/{planId}:
    get:
      summary: Find schedule linked to a VTMS test plan
      parameters:
        - in: path
          name: planId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Linked schedule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '404':
          description: No schedule linked to this plan
        '401':
          description: Unauthorized

  /schedules-with-progress:
    get:
      summary: Schedules with VTMS progress data (Copilot cross-system query)
      parameters:
        - { in: query, name: testUnit,     schema: { type: string } }
        - { in: query, name: testEngineer, schema: { type: string }, description: 'Exact test engineer name' }
        - { in: query, name: device,       schema: { type: string }, description: 'Exact equipment/device name' }
        - { in: query, name: projectName,  schema: { type: string }, description: 'Partial project name match' }
        - { in: query, name: isCompleted,  schema: { type: string, enum: ['true','false'] } }
        - { in: query, name: isDelayed,    schema: { type: string, enum: ['true','false'] } }
        - { in: query, name: isCancelled,  schema: { type: string, enum: ['true','false'] } }
        - { in: query, name: userFlag,     schema: { type: string, enum: ['true','false'] }, description: 'Filter by user (non-admin) mark' }
        - { in: query, name: dateFrom,     schema: { type: string } }
        - { in: query, name: dateTo,       schema: { type: string } }
      responses:
        '200':
          description: Schedules with embedded VTMS progress (null if not linked or VTMS unavailable)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduleWithProgress'
        '401':
          description: Unauthorized

  /workload-analysis:
    get:
      summary: Engineer workload analysis for one month (algorithm runs server-side; cancelled schedules excluded)
      parameters:
        - { in: query, name: month,        required: true, schema: { type: string }, description: 'Target month, YYYY-MM' }
        - { in: query, name: testEngineer, schema: { type: string }, description: 'Exact test engineer name' }
        - { in: query, name: testUnit,     schema: { type: string }, description: 'Exact test unit name' }
        - { in: query, name: overtime,     schema: { type: string }, description: 'Overtime hours per engineer, "Name1=8,Name2=4.5" (from Work IQ)' }
      responses:
        '200':
          description: Per-engineer workload sorted by rate desc
          content:
            application/json:
              schema:
                type: object
                properties:
                  month:    { type: string }
                  workdays: { type: integer, description: 'Business days in month (weekends and imported holidays excluded)' }
                  limitations:
                    type: array
                    items: { type: string }
                  engineers:
                    type: array
                    items:
                      type: object
                      properties:
                        testEngineer:    { type: string }
                        testUnits:       { type: array, items: { type: string } }
                        scheduleCount:   { type: integer }
                        baseScore:       { type: number, description: 'Σ min(daily intensity sum, 1.2) over month workdays' }
                        overtimeHours:   { type: number, nullable: true }
                        overtimeBonus:   { type: number, nullable: true, description: 'hours ÷ 6 (6h = 1 point), uncapped' }
                        total:           { type: number }
                        rate:            { type: number, description: 'total ÷ month workdays × 100 (%)' }
                        level:           { type: string, enum: ['超載', '滿載', '中等', '偏低'] }
                        unscheduledDays: { type: integer }
                        partialDays:     { type: integer, description: 'Workdays with pre-cap intensity < 1' }
                        cappedDays:      { type: integer, description: 'Workdays with pre-cap intensity ≥ 1.2' }
                        limitations:     { type: array, items: { type: string } }
        '400':
          description: Missing or malformed month
        '401':
          description: Unauthorized

  /schedules/{id}/delay:
    patch:
      summary: Mark schedule as delayed (called by VTMS on Delay log)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [date, content]
              properties:
                date:    { type: string, description: 'YYYY-MM-DD' }
                content: { type: string }
      responses:
        '200':
          description: Updated schedule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '400':
          description: Missing date or content
        '404':
          description: Schedule not found
        '401':
          description: Unauthorized

  /schedules/{id}/complete:
    patch:
      summary: Mark schedule as completed (called by VTMS when all tasks done)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Updated schedule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '404':
          description: Schedule not found
        '401':
          description: Unauthorized